build dir

+++ irprimer.py
@@ -2,7 +2,7 @@ import os
 from random import random
 import time
 
-from subprocess import Popen
+from subprocess import Popen, PIPE
 from IPython.display import Javascript, HTML
 
 replacements = {
@@ -13,38 +13,34 @@ replacements = {
     "https://cdn.mathjax.org/mathjax/latest/":
         "./bower_components/MathJax/",
     "https://cdnjs.cloudflare.com/ajax/libs/require.js/2.1.10/require.min.js":
-        "./bower_components/requirejs/require.js"
+        "./bower_components/requirejs/require.js",
+    "./dist/": "./",
+    '<link rel="stylesheet" href="custom.css">': ""
 }
 
 
 def slides(infile="An IR Primer.ipynb",
-           tmpfile="An IR Primer.slides.html",
-           outfile="An IR Primer.slides.local.html"):
-    Javascript("""
-        Reveal.addEventListener( 'slidechanged', function( event ) {
-            window.resize()
-        });
-        IPython.notebook.save_notebook()
-    """)
+           outfile="dist/An IR Primer.slides.local.html"):
+    Javascript("IPython.notebook.save_notebook()")
 
     time.sleep(1)
 
-    Popen([
+    html, err = Popen([
         "ipython", "nbconvert",
         infile,
         "--to=slides",
-        "--reveal-prefix=./bower_components/reveal.js"
-    ])
+        "--reveal-prefix=./bower_components/reveal.js",
+        "--stdout"
+    ], stdout=PIPE).communicate()
 
-    time.sleep(1)
-
-    with open(tmpfile, "r") as f:
-        html = f.read()
+    html = html.decode("utf-8")
 
     for remote, local in replacements.items():
         html = html.replace(remote, local)
 
-    os.unlink(outfile)
+    if os.path.exists(outfile):
+        os.unlink(outfile)
+
     with open(outfile, "w+") as f:
         f.write(html)
 
@@ -52,30 +48,43 @@ def slides(infile="An IR Primer.ipynb",
 def planck():
     return HTML("""
         <style>
-            @import url("planck.css");
-            #planck-{id}{{
+            @import url("./dist/planck.css");
+            #planck-%(id)s{
                 background-color: white;
-
-            }}
-            .reveal #planck-{id} {{
-                margin-top: -200px;
-                margin-left: -300px;
-            }}
+            }
         </style>
-        <div id="planck-{id}" class="stretch"></div>
+        <div id="planck-%(id)s" class="stretch"></div>
         <script>
-            require(["./bower_components/d3/d3.min.js", "./planck.js"],
-                function(d3, Planck){{
+            require(
+                [
+                    "./dist/bower_components/d3/d3.min.js",
+                    "./dist/planck.js"
+                ],
+                function(d3, Planck){
                     var planck = Planck(d3);
-                    d3.select("#planck-{id}")
-                        .style({{
+                    var wrapper = d3.select("#planck-%(id)s")
+                        .style({
                             "background-color": "white",
-                            width: window.innerWidth * 0.9 + "px",
-                            height: window.innerHeight * 0.9 + "px",
-                        }})
-                        .call(planck)
-                }});
+                            width: "100%%",
+                            height: "800px",
+                        })
+                        .call(planck);
+                    if(window.Reveal){
+                        var _window = d3.select(window);
+                        Reveal.addEventListener(
+                            "slidechanged",
+                            function(event){
+                                $(window).resize();
+                            }
+                        );
+                        .on("resize.planck-%(id)s", function(){
+                            wrapper.style({
+                                width: window.innerWidth + "px",
+                                height: window.innerHeight + "px"
+                            })
+                            planck.resize();
+                        });
+                    }
+            });
         </script>
-
-    """.format(id=str(random()).split(".")[1])
-    )
+        """ % dict(id=str(random()).split(".")[1]))
